refactor(test): extract parseCommandFrontmatter helper #656
Move inline frontmatter parsing logic from pm-review-fix-jest.test.js
into test/helpers/parse-command-frontmatter.js for reuse in future
command-file tests.
